do not use GtkApplication
The architecture with GtkApplication is to have a single process with
several windows. When a user is starting a new instance all it does is
tell the existing process to add new window.

Since I did not know that the current implementation is buggy:

-	When one window is closed, the whole application quits and all of its
	windows are closed along with it.
-	Additional windows are not raised on creation.

To fix this I would have to change the code to not use the global window
and notebook variables.

As far as I could find out, the "old" approach to have one process per
instance is not deprecated and just simpler to reason about. So I prefer
to do that.

https://developer.gnome.org/GtkApplication/
https://wiki.gnome.org/Projects/GnomeShell/ApplicationBased